feat(employee): implement complete end-to-end employee management and onboarding module
This commit introduces the full lifecycle for Employee registration, including dynamic UI forms, complex backend relations, automated user account provisioning, structured file storage, and async email notifications.

Frontend & UI:
- Designed a premium, dense slate-gray/gold "Add Employee" form with collapsible accordions.
- Implemented dynamic row additions for Contacts, Qualifications, Experiences, and Documents.
- Built complex payload orchestration in employees.js to parse form data and Base64 encode files.
- Added custom premium success modals for user feedback.

Backend Core & Architecture:
- Implemented EmployeeController, EmployeeServiceImpl, and ErpEmployeeRepository.
- Added deep JPA one-to-many cascading relationships for ErpEmployeeContact, ErpEmployeeQualification, ErpEmployeeExperience, and ErpEmployeeDocument.
- Created robust DTO mappers to handle nested JSON collections cleanly.
- Implemented automatic human-readable employeeNo generation.

Storage & File Management:
- Restructured StorageService.storeEntityDocument to use human-readable, dynamic paths.
- Consolidated employee uploads into a single directory format: uploads/{schoolCode-name,location}/staff/{employeeNo-name}/.
- Added automatic collision-prevention prefixes (e.g., qual_, exp_) to uploaded files.

Security & Automated Provisioning:
- Integrated UserService and PasswordService directly into the Employee registration flow.
- Added automated ErpUser account creation mapped directly to the new employee.
- Resolved dynamic default role assignment using ErpRoleRepository.
- Enclosed employee and user creation within a strict @Transactional boundary to prevent orphaned records.

Email & Notifications:
- Implemented asynchronous (@Async) welcome emails via EmailService.
- Created a professional Thymeleaf email template (employee-welcome.html) sending the generated username and temporary secure password.

Fixes & Optimizations:
- Resolved ByteBuddyInterceptor serialization errors during JSON responses.
- Fixed UI null pointer exceptions on optional middle names and dynamically added rows.
- Refactored payload structures to ensure perfect sync between JS payload and Java DTOs.

+++ montfortuganda/src/main/java/com/erp/montfortuganda/auth/repository/ErpRoleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ErpRoleRepository extends JpaRepository<ErpRole, Integer> {
+public interface ErpRoleRepository extends JpaRepository<ErpRole, Long> {
     Optional<ErpRole> findByRoleCode(String roleCode);
 }
